Cleanup and rearrange tests
