+++ tests/components/jewish_calendar/test_sensor.py
@@ -1,9 +1,6 @@
-"""The tests for the Jewish calendar sensor platform."""
-from collections import namedtuple
-from contextlib import contextmanager
+"""The tests for the Jewish calendar sensors."""
 from datetime import time
 from datetime import datetime as dt
-from unittest.mock import patch
 
 import pytest
 
@@ -11,49 +8,7 @@ from homeassistant.util.dt import get_time_zone
 from homeassistant.setup import async_setup_component
 from homeassistant.components import jewish_calendar
 
-
-_LatLng = namedtuple("_LatLng", ["lat", "lng"])
-
-NYC_LATLNG = _LatLng(40.7128, -74.0060)
-JERUSALEM_LATLNG = _LatLng(31.778, 35.235)
-
-
-def make_nyc_test_params(dtime, results, havdalah_offset=0):
-    """Make test params for NYC."""
-    return (
-        dtime,
-        jewish_calendar.CANDLE_LIGHT_DEFAULT,
-        havdalah_offset,
-        True,
-        "America/New_York",
-        NYC_LATLNG.lat,
-        NYC_LATLNG.lng,
-        results,
-    )
-
-
-def make_jerusalem_test_params(dtime, results, havdalah_offset=0):
-    """Make test params for Jerusalem."""
-    return (
-        dtime,
-        jewish_calendar.CANDLE_LIGHT_DEFAULT,
-        havdalah_offset,
-        False,
-        "Asia/Jerusalem",
-        JERUSALEM_LATLNG.lat,
-        JERUSALEM_LATLNG.lng,
-        results,
-    )
-
-
-@contextmanager
-def alter_time(retval):
-    """Manage multiple time mocks."""
-    patch1 = patch("homeassistant.util.dt.utcnow", return_value=retval)
-    patch2 = patch("homeassistant.util.dt.now", return_value=retval)
-
-    with patch1, patch2:
-        yield
+from . import alter_time, make_nyc_test_params, make_jerusalem_test_params
 
 
 class TestJewishCalenderSensor:
@@ -597,111 +552,19 @@ class TestJewishCalenderSensor:
             ), f"Value for {sensor_type}"
         hass.config.set_time_zone(default_time_zone)
 
-    melacha_params = [
-        make_nyc_test_params(dt(2018, 9, 1, 16, 0), "on"),
-        make_nyc_test_params(dt(2018, 9, 1, 20, 21), "off"),
-        make_nyc_test_params(dt(2018, 9, 7, 13, 1), "off"),
-        make_nyc_test_params(dt(2018, 9, 8, 21, 25), "off"),
-        make_nyc_test_params(dt(2018, 9, 9, 21, 25), "on"),
-        make_nyc_test_params(dt(2018, 9, 10, 21, 25), "on"),
-        make_nyc_test_params(dt(2018, 9, 28, 21, 25), "on"),
-        make_nyc_test_params(dt(2018, 9, 29, 21, 25), "off"),
-        make_nyc_test_params(dt(2018, 9, 30, 21, 25), "on"),
-        make_nyc_test_params(dt(2018, 10, 1, 21, 25), "on"),
-        make_jerusalem_test_params(dt(2018, 9, 29, 21, 25), "off"),
-        make_jerusalem_test_params(dt(2018, 9, 30, 21, 25), "on"),
-        make_jerusalem_test_params(dt(2018, 10, 1, 21, 25), "off"),
-    ]
-    melacha_test_ids = [
-        "currently_first_shabbat",
-        "after_first_shabbat",
-        "friday_upcoming_shabbat",
-        "upcoming_rosh_hashana",
-        "currently_rosh_hashana",
-        "second_day_rosh_hashana",
-        "currently_shabbat_chol_hamoed",
-        "upcoming_two_day_yomtov_in_diaspora",
-        "currently_first_day_of_two_day_yomtov_in_diaspora",
-        "currently_second_day_of_two_day_yomtov_in_diaspora",
-        "upcoming_one_day_yom_tov_in_israel",
-        "currently_one_day_yom_tov_in_israel",
-        "after_one_day_yom_tov_in_israel",
-    ]
-
-    @pytest.mark.parametrize(
-        [
-            "now",
-            "candle_lighting",
-            "havdalah",
-            "diaspora",
-            "tzname",
-            "latitude",
-            "longitude",
-            "result",
-        ],
-        melacha_params,
-        ids=melacha_test_ids,
-    )
-    async def test_issur_melacha_sensor(
-        self,
-        hass,
-        now,
-        candle_lighting,
-        havdalah,
-        diaspora,
-        tzname,
-        latitude,
-        longitude,
-        result,
-    ):
-        """Test Issur Melacha sensor output."""
-        time_zone = get_time_zone(tzname)
-        test_time = time_zone.localize(now)
-
-        default_time_zone = str(hass.config.time_zone)
-        hass.config.set_time_zone(tzname)
-        hass.config.latitude = latitude
-        hass.config.longitude = longitude
-
-        assert await async_setup_component(
-            hass,
-            jewish_calendar.DOMAIN,
-            {
-                "jewish_calendar": {
-                    "name": "test",
-                    "language": "english",
-                    "diaspora": diaspora,
-                    "candle_lighting_minutes_before_sunset": candle_lighting,
-                    "havdalah_minutes_after_sunset": havdalah,
-                }
-            },
-        )
-        await hass.async_block_till_done()
-
-        with alter_time(test_time):
-            await hass.helpers.entity_component.async_update_entity(
-                "binary_sensor.test_issur_melacha_in_effect"
-            )
-
-        assert (
-            hass.states.get("binary_sensor.test_issur_melacha_in_effect").state
-            == result
-        )
-        hass.config.set_time_zone(default_time_zone)
-
     omer_params = [
-        make_nyc_test_params(dt(2019, 4, 21, 0, 0), 1),
-        make_jerusalem_test_params(dt(2019, 4, 21, 0, 0), 1),
-        make_nyc_test_params(dt(2019, 4, 21, 23, 0), 2),
-        make_jerusalem_test_params(dt(2019, 4, 21, 23, 0), 2),
-        make_nyc_test_params(dt(2019, 5, 23, 0, 0), 33),
-        make_jerusalem_test_params(dt(2019, 5, 23, 0, 0), 33),
-        make_nyc_test_params(dt(2019, 6, 8, 0, 0), 49),
-        make_jerusalem_test_params(dt(2019, 6, 8, 0, 0), 49),
-        make_nyc_test_params(dt(2019, 6, 9, 0, 0), 0),
-        make_jerusalem_test_params(dt(2019, 6, 9, 0, 0), 0),
-        make_nyc_test_params(dt(2019, 1, 1, 0, 0), 0),
-        make_jerusalem_test_params(dt(2019, 1, 1, 0, 0), 0),
+        make_nyc_test_params(dt(2019, 4, 21, 0, 0), "1"),
+        make_jerusalem_test_params(dt(2019, 4, 21, 0, 0), "1"),
+        make_nyc_test_params(dt(2019, 4, 21, 23, 0), "2"),
+        make_jerusalem_test_params(dt(2019, 4, 21, 23, 0), "2"),
+        make_nyc_test_params(dt(2019, 5, 23, 0, 0), "33"),
+        make_jerusalem_test_params(dt(2019, 5, 23, 0, 0), "33"),
+        make_nyc_test_params(dt(2019, 6, 8, 0, 0), "49"),
+        make_jerusalem_test_params(dt(2019, 6, 8, 0, 0), "49"),
+        make_nyc_test_params(dt(2019, 6, 9, 0, 0), "0"),
+        make_jerusalem_test_params(dt(2019, 6, 9, 0, 0), "0"),
+        make_nyc_test_params(dt(2019, 1, 1, 0, 0), "0"),
+        make_jerusalem_test_params(dt(2019, 1, 1, 0, 0), "0"),
     ]
     omer_test_ids = [
         "nyc_first_day_of_omer",
@@ -773,5 +636,5 @@ class TestJewishCalenderSensor:
                 "sensor.test_day_of_the_omer"
             )
 
-        assert hass.states.get("sensor.test_day_of_the_omer").state == str(result)
+        assert hass.states.get("sensor.test_day_of_the_omer").state == result
         hass.config.set_time_zone(default_time_zone)
